+++ src/main/java/net/abraxator/moresnifferflowers/init/ModRecipeSerializers.java
@@ -1,8 +1,8 @@
 package net.abraxator.moresnifferflowers.init;
 
 import net.abraxator.moresnifferflowers.MoreSnifferFlowers;
+import net.abraxator.moresnifferflowers.recipes.CropressingRecipe;
 import net.abraxator.moresnifferflowers.recipes.RebrewedTippedArrowRecipe;
-import net.abraxator.moresnifferflowers.recipes.serializers.CropressingSerializer;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.world.item.crafting.Recipe;
 import net.minecraft.world.item.crafting.RecipeSerializer;
@@ -14,6 +14,6 @@ public class ModRecipeSerializers {
     public static final DeferredRegister<RecipeSerializer<?>> RECIPE_SERIALIZERS =
             DeferredRegister.create(BuiltInRegistries.RECIPE_SERIALIZER, MoreSnifferFlowers.MOD_ID);
 
-    public static final DeferredHolder<RecipeSerializer<?>, CropressingSerializer> CROPRESSING = RECIPE_SERIALIZERS.register("cropressing", CropressingSerializer::new);
+    public static final DeferredHolder<RecipeSerializer<?>, CropressingRecipe.CropressingSerializer> CROPRESSING = RECIPE_SERIALIZERS.register("cropressing", CropressingRecipe.CropressingSerializer::new);
     public static final DeferredHolder<RecipeSerializer<?>, RecipeSerializer<RebrewedTippedArrowRecipe>> REBREWED_TIPPED_ARROW = RECIPE_SERIALIZERS.register("rebrewed_tipped_arrow", () -> new SimpleCraftingRecipeSerializer<RebrewedTippedArrowRecipe>(RebrewedTippedArrowRecipe::new));
 }